refactor(inference): extract qnn metrics tests to sibling file
`primer-inference/src/qnn/metrics.rs` carried a single inline trailing
`#[cfg(test)] mod tests` (metrics-path resolution, rotation, metered-stream
record-once / finalize-on-error / break-on-done, and append/rotate file I/O).
Move it verbatim into a sibling `src/qnn/metrics/tests.rs`. metrics.rs is a
non-root module, so the submodule resolves under `src/qnn/metrics/` (mkdir).
The test module is a descendant, so `use super::*` reaches all parent privates
with no visibility edits — zero churn.

508 -> 271 lines (-237). All 11 qnn::metrics tests preserved; `--features qnn`
lib suite still 281 passed / 0 failed. Pure test-only relocation, no behaviour
change.
